Create LazyEncodingRegistryTest for testing LazyEncodingRegistry
- Create LazyEncodingRegistryTest for testing LazyEncodingRegistry.
- Refactor tests, create BaseEncodingRegistryTest for reusing tests for LazyEncodingRegistryTest and DefaultEncodingRegistryTest.
